+++ config/config.yaml
@@ -32,6 +32,8 @@ app:
   # 【本系统的WEB的外网地址】：需要是外网IP或者域名，需要包含端口，用于微信/Telegram信息点击跳转，如不需要可配空
   # 示例：http://IP:3000
   domain: ''
+  # 【UserAgent】：可适当修改，用于PT站签到、豆瓣数据抓取等
+  user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/98.0.4758.102 Safari/537.36"
 
 # 【配置媒体库信息】
 media:
@@ -180,8 +182,6 @@ pt:
   # 【PT站每日签到时间】
   # 两种配置方法，1、配置间隔，单位小时，建议不要设置为24小时的整数倍，避免每天的签到时间一样。2、配置固定时间，如'08:00'，注意要加引号和冒号
   ptsignin_cron: '08:01'
-  # 【PT站签到UserAgent】：可适当修改，以防止被识别为机器人签到
-  user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/98.0.4758.102 Safari/537.36"
   # 【PT做种清理开关】：此处配置PT做种时间，做种超过该时间则会自动删种（包括已下载的文件），单位是【天】。
   # 如果是使用COPY模式，这个根据PT下载盘的大小配置，如果是使用LINK模式，这个一般不配置，
   # 不配置或者配置为0则不启用PT做种清理功能
@@ -283,5 +283,18 @@ douban:
   auto_search: true
   # 【自动添加RSS开关】：PT站检索找不到的记录是否自动添加RSS订阅（可实现未搜索到的自动追更）
   auto_rss: true
-  # 【UserAgent】：建议填写，以免触发触发豆瓣的反爬虫机制浏览器获取
-  user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/98.0.4758.102 Safari/537.36"
+
+# 【配置字幕自动下载】
+subtitle:
+  # 【下载渠道】：opensubtitles、chinesesubfinder
+  server: opensubtitles
+  # 配置OpenSubtitles.org中注册的用户名和密码
+  opensubtitles:
+    # 用户名
+    username:
+    # 密码
+    password:
+  # 配置ChineseSubFinder的服务器地址和API Key，API Key在ChineseSubFinder->配置中心->实验室->API Key处生成
+  chinesesubfinder:
+    host:
+    api_key:
