+++ pt/subtitle.py
@@ -1,0 +1,11 @@
+
+class Subtitle:
+
+    def __init__(self):
+        self.init_config()
+
+    def init_config(self):
+        pass
+
+    def download_subtitle(self):
+        pass
