+++ test/run.py
@@ -1,18 +1,4 @@
-from pt.client.qbittorrent import Qbittorrent
-from pt.searcher import Searcher
-from rmt.filetransfer import FileTransfer
 from rmt.media import Media
-from rmt.metainfo import MetaInfo
-from rmt.server.jellyfin import Jellyfin
-from rmt.server.plex import Plex
-from scheduler.douban_sync import DoubanSync
-from scheduler.rss_search import RssSearch
-from scheduler.scheduler import Scheduler
-from utils.db_helper import select_by_sql, update_by_sql
-from utils.sqls import get_config_site, get_config_search_rule, insert_transfer_unknown, insert_rss_movie, \
-    insert_rss_tv, get_rss_tvs, get_transfer_statistics
-from utils.types import SyncType, SearchType
-from web.backend.subscribe import add_rss_substribe_from_string
 
 if __name__ == "__main__":
     '''
