+++ web/templates/navigation.html
@@ -230,6 +230,9 @@
                   <a class="dropdown-item" href="javascript:navmenu('library')">
                     媒体库
                   </a>
+                  <a class="dropdown-item" href="javascript:navmenu('directorysync')">
+                    目录同步
+                  </a>
                   <a class="dropdown-item" href="javascript:navmenu('notification')">
                     消息通知
                   </a>
@@ -242,8 +245,8 @@
                   <a class="dropdown-item" href="javascript:navmenu('mediaserver')">
                     媒体服务器
                   </a>
-                  <a class="dropdown-item" href="javascript:navmenu('directorysync')">
-                    目录同步
+                  <a class="dropdown-item" href="javascript:navmenu('subtitle')">
+                    字幕
                   </a>
                   <a class="dropdown-item" href="javascript:navmenu('douban')">
                     豆瓣
